+++ src/main/java/org/example/demo/router/Router.java
@@ -1,8 +1,9 @@
-package org.example.demo;
+package org.example.demo.router;
 
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
+import org.example.demo.HttpMethod;
 import org.example.demo.exception.InternalServerError;
 import org.example.demo.exception.UnauthorizedException;
 import org.slf4j.Logger;
