+++ src/main/java/org/example/demo/servlet/UserServlet.java
@@ -5,9 +5,8 @@ import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import org.example.demo.HttpMethod;
-import org.example.demo.Router;
+import org.example.demo.router.Router;
 import org.example.demo.exception.InternalServerError;
-import org.example.demo.exception.UnauthorizedException;
 import org.example.demo.handler.UserHandler;
 
 import java.io.IOException;
